# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Goto Backend",
    "description": "Goto 后端服务：提供任务、项目、分类、标签的 CRUD 管理，供桌面端通过 HTTP API 调用。与端到端加密同步链路解耦，仅作可选远程存储。",
    "contact": {
      "name": "Goto Security",
      "email": "security@ms33834.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tasks/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "获取任务列表",
        "description": "获取任务列表，支持按状态/优先级/分类/项目/完成态过滤与排序。\n\n不传 sort_by 时按默认顺序 (order, created_at desc) 排序。",
        "operationId": "list_tasks_api_v1_tasks__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Deleted"
            }
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          },
          {
            "name": "completed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按完成状态过滤",
              "title": "Completed"
            },
            "description": "按完成状态过滤"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态过滤",
              "title": "Status"
            },
            "description": "按状态过滤"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按优先级过滤",
              "title": "Priority"
            },
            "description": "按优先级过滤"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按分类 ID 过滤",
              "title": "Category Id"
            },
            "description": "按分类 ID 过滤"
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按项目 ID 过滤",
              "title": "Project Id"
            },
            "description": "按项目 ID 过滤"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "排序字段: created_at / updated_at / due_date / priority / title / order",
              "title": "Sort By"
            },
            "description": "排序字段: created_at / updated_at / due_date / priority / title / order"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "排序方向: asc / desc",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "排序方向: asc / desc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Tasks Api V1 Tasks  Get"
                }
              }
            }
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "422": {
            "description": "参数校验失败"
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "创建任务",
        "description": "创建任务",
        "operationId": "create_task_api_v1_tasks__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "422": {
            "description": "参数校验失败"
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "获取任务详情",
        "description": "获取单个任务",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "422": {
            "description": "参数校验失败"
          },
          "404": {
            "description": "资源不存在"
          }
        }
      },
      "patch": {
        "tags": [
          "tasks"
        ],
        "summary": "更新任务",
        "description": "更新任务\n\n乐观锁采用单条原子 UPDATE 实现，避免 TOCTOU 竞态（H-2）：\nUPDATE tasks SET ..., version = version + 1, updated_at = :now\nWHERE id = :task_id AND version = :expected_version\n通过 result.rowcount 判断是否命中：\n- rowcount == 1：更新成功\n- rowcount == 0：任务不存在或 version 不匹配，二次 SELECT 区分 404 / 409\n不传 expected_version 时 WHERE 仅含 id 条件，保持向后兼容。",
        "operationId": "update_task_api_v1_tasks__task_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "422": {
            "description": "参数校验失败"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突（乐观锁）"
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "删除任务",
        "description": "删除任务（默认软删除）",
        "operationId": "delete_task_api_v1_tasks__task_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "hard",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Hard"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "404": {
            "description": "资源不存在"
          },
          "422": {
            "description": "参数校验失败"
          }
        }
      }
    },
    "/api/v1/tasks/projects/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "获取项目列表",
        "description": "获取项目列表",
        "operationId": "list_projects_api_v1_tasks_projects__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "type": "array",
                  "title": "Response List Projects Api V1 Tasks Projects  Get"
                }
              }
            }
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "422": {
            "description": "参数校验失败"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "创建项目",
        "description": "创建项目",
        "operationId": "create_project_api_v1_tasks_projects__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "422": {
            "description": "参数校验失败"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/tasks/categories/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "获取分类列表",
        "operationId": "list_categories_api_v1_tasks_categories__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  },
                  "type": "array",
                  "title": "Response List Categories Api V1 Tasks Categories  Get"
                }
              }
            }
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "422": {
            "description": "参数校验失败"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "创建分类",
        "operationId": "create_category_api_v1_tasks_categories__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "422": {
            "description": "参数校验失败"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "color": {
            "type": "string",
            "title": "Color",
            "default": "#6b7280"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "is_system": {
            "type": "boolean",
            "title": "Is System",
            "default": false
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived",
            "default": false
          },
          "parent_category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Category Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "order": {
            "type": "integer",
            "title": "Order",
            "default": 0
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CategoryCreate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}